refactor(tui): move message history renderer
Extract user, assistant, and system message rendering plus copy metadata helpers from history.rs into history/message.rs while preserving the existing history API and test surface.

Refs #3308.
